+++ py-xtb/pyproject.toml
@@ -1,0 +1,11 @@
+[project]
+name = "py-xtb"
+version = "0.1.0"
+description = "Add your description here"
+readme = "README.md"
+requires-python = ">=3.11"
+dependencies = []
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
