+++ py-xtb/src/py_xtb/__init__.py
@@ -1,0 +1,1 @@
+from .geometry import Atom, Geometry
