+++ py-xtb/src/py_xtb/freq.py
@@ -1,46 +1,47 @@
 # Copyright (c) 2023-2024, Matthew J. Milner
-# This file is part of avo_xtb which is released under the BSD 3-Clause License.
+# This file is part of py-xtb which is released under the BSD 3-Clause License.
 # See LICENSE or go to https://opensource.org/license/BSD-3-clause for full details.
 
 import argparse
 import json
 import sys
-from shutil import rmtree
 from pathlib import Path
+from shutil import rmtree
 
 from config import config, calc_dir
+import obabel_convert
 from run import run_xtb
 
 
-def orbitals(
+def frequencies(
     geom_file: Path,
     charge: int = 0,
     multiplicity: int = 1,
     solvation: str | None = None,
     method: int = 2,
 ) -> Path:
-    """Calculate molecular orbitals for given geometry, return file in Molden format."""
-    spin = multiplicity - 1
-    # Just do a single point calculation but pass molden option to get orbital printout
+    """Calculate vibrational frequencies and return Gaussian 98 format output file."""
+    unpaired_e = multiplicity - 1
     command = [
         "xtb",
         geom_file,
-        "--molden",
+        "--hess",
         "--chrg",
         str(charge),
         "--uhf",
-        str(spin),
+        str(unpaired_e),
         "--gfn",
         str(method),
     ]
+
     # Add solvation if requested
     if solvation is not None:
         command.extend(["--alpb", solvation])
     # Run xtb from command line
     calc, out_file, energy = run_xtb(command, geom_file)
 
-    # Return path to molden file
-    return geom_file.with_name("molden.input")
+    # Return the path of the Gaussian file generated
+    return geom_file.with_name("g98.out")
 
 
 if __name__ == "__main__":
@@ -57,9 +58,9 @@ if __name__ == "__main__":
         options = {"inputMoleculeFormat": "xyz"}
         print(json.dumps(options))
     if args.display_name:
-        print("Orbitals")
+        print("Frequencies")
     if args.menu_path:
-        print("Extensions|Semi-empirical (xtb){830}")
+        print("Extensions|Semi-empirical (xtb){870}")
 
     if args.run_command:
         # Remove results of last calculation
@@ -74,12 +75,12 @@ if __name__ == "__main__":
         avo_input = json.loads(sys.stdin.read())
         # Extract the coords and write to file for use as xtb input
         geom = avo_input["xyz"]
-        xyz_path = calc_dir / "input.xyz"
+        xyz_path = Path(calc_dir) / "input.xyz"
         with open(xyz_path, "w", encoding="utf-8") as xyz_file:
             xyz_file.write(str(geom))
 
-        # Run calculation using xyz file
-        result_path = orbitals(
+        # Run calculation; returns path to Gaussian file containing frequencies
+        result_path = frequencies(
             xyz_path,
             charge=avo_input["charge"],
             multiplicity=avo_input["spin"],
@@ -87,26 +88,32 @@ if __name__ == "__main__":
             method=config["method"],
         )
 
-        # Get molden file as string
-        with open(result_path, encoding="utf-8") as molden_file:
-            molden_string = molden_file.read()
-        # Format everything appropriately for Avogadro
-        # Just pass orbitals file with instruction to read only properties
-        result = {
-            "readProperties": True,
-            "moleculeFormat": "molden",
-            "molden": molden_string,
-        }
-        # As it stands, this means any other properties will be wiped
-        # If there were e.g. frequencies in the original cjson, notify the user
-        if "vibrations" in avo_input["cjson"]:
-            result["message"] = (
-                "Calculation complete!\n"
-                "The vibrational frequencies may have been lost in this process.\n"
-                "Please recalculate them if they are missing and still desired.\n"
-            )
-        else:
-            result["message"] = "Calculation complete!"
+        # Currently Avogadro fails to convert the g98 file to cjson itself
+        # So we have to convert output in g98 format to cjson ourselves
+        cjson_path = obabel_convert.g98_to_cjson(result_path)
+        # Open the cjson
+        with open(cjson_path, encoding="utf-8") as result_cjson:
+            freq_cjson = json.load(result_cjson)
+        # Format appropriately for Avogadro
+        # Start by passing back the original cjson, then add changes
+        result = {"moleculeFormat": "cjson", "cjson": avo_input["cjson"]}
+        result["cjson"]["vibrations"] = freq_cjson["vibrations"]
+        # xtb no longer gives Raman intensities but does write them as all 0
+        # If passed on to the user this is misleading so remove them
+        if "ramanIntensities" in result["cjson"]["vibrations"]:
+            del result["cjson"]["vibrations"]["ramanIntensities"]
 
+        # Inform user if there are negative frequencies
+        if float(freq_cjson["vibrations"]["frequencies"][0]) < 0:
+            result["message"] = (
+                "At least one negative frequency found!\n"
+                + "This is not a minimum on the potential energy surface.\n"
+                + "You should reoptimize the geometry.\n"
+                + "This can be avoided in future by using the Opt + Freq method."
+            )
+
+        # Save result
+        with open(calc_dir / "result.cjson", "w", encoding="utf-8") as save_file:
+            json.dump(result["cjson"], save_file, indent=2)
         # Pass back to Avogadro
         print(json.dumps(result))
