+++ py-xtb/src/py_xtb/obabel_convert.py
@@ -1,5 +1,5 @@
 # Copyright (c) 2023-2024, Matthew J. Milner
-# This file is part of avo_xtb which is released under the BSD 3-Clause License.
+# This file is part of py-xtb which is released under the BSD 3-Clause License.
 # See LICENSE or go to https://opensource.org/license/BSD-3-clause for full details.
 
 """Converts files to another format using an Open Babel binary."""
