+++ py-xtb/src/py_xtb/opt.py
@@ -1,30 +1,33 @@
 # Copyright (c) 2023-2024, Matthew J. Milner
-# This file is part of avo_xtb which is released under the BSD 3-Clause License.
+# This file is part of py-xtb which is released under the BSD 3-Clause License.
 # See LICENSE or go to https://opensource.org/license/BSD-3-clause for full details.
 
 import argparse
 import json
 import sys
-from pathlib import Path
 from shutil import rmtree
+from pathlib import Path
 
 from config import config, calc_dir
-from run import run_xtb
 import convert
+from run import run_xtb
 
 
-def energy(
+def optimize(
     geom_file: Path,
     charge: int = 0,
     multiplicity: int = 1,
     solvation: str | None = None,
     method: int = 2,
-) -> float:
-    """Calculate energy in hartree for given geometry."""
+    level: str = "normal",
+) -> tuple[Path, float]:
+    """Return optimized geometry as file in same format as the input, along with the energy."""
     unpaired_e = multiplicity - 1
     command = [
         "xtb",
         geom_file,
+        "--opt",
+        level,
         "--chrg",
         str(charge),
         "--uhf",
@@ -32,12 +35,14 @@ def energy(
         "--gfn",
         str(method),
     ]
+
     # Add solvation if requested
     if solvation is not None:
         command.extend(["--alpb", solvation])
     # Run xtb from command line
     calc, out_file, energy = run_xtb(command, geom_file)
-    return energy
+    # Return path to geometry file with same suffix, along with energy
+    return geom_file.with_stem("xtbopt"), energy
 
 
 if __name__ == "__main__":
@@ -54,9 +59,9 @@ if __name__ == "__main__":
         options = {"inputMoleculeFormat": "xyz"}
         print(json.dumps(options))
     if args.display_name:
-        print("Energy")
+        print("Optimize")
     if args.menu_path:
-        print("Extensions|Semi-empirical (xtb){890}")
+        print("Extensions|Semi-empirical (xtb){880}")
 
     if args.run_command:
         # Remove results of last calculation
@@ -71,31 +76,43 @@ if __name__ == "__main__":
         avo_input = json.loads(sys.stdin.read())
         # Extract the coords and write to file for use as xtb input
         geom = avo_input["xyz"]
-        xyz_path = Path(calc_dir) / "input.xyz"
+        xyz_path = calc_dir / "input.xyz"
         with open(xyz_path, "w", encoding="utf-8") as xyz_file:
             xyz_file.write(str(geom))
 
-        # Run calculation; returns energy as float in hartree
-        energy_hartree = energy(
+        # Run calculation using xyz file
+        result_path, energy = optimize(
             xyz_path,
             charge=avo_input["charge"],
             multiplicity=avo_input["spin"],
             solvation=config["solvent"],
             method=config["method"],
+            level=config["opt_lvl"],
         )
-        # Convert energy to eV for Avogadro, other units for users
-        energies = convert.convert_energy(energy_hartree, "hartree")
+
+        # Read the xyz file
+        with open(result_path.with_name("xtbopt.xyz"), encoding="utf-8") as result_xyz:
+            xyz = result_xyz.read().split("\n")
+        # Convert geometry
+        cjson_geom = convert.xyz_to_cjson(xyz_lines=xyz)
+        # Check for convergence
+        # TODO
+        # Will need to look for "FAILED TO CONVERGE"
+        # Convert energy for Avogadro
+        energies = convert.convert_energy(energy, "hartree")
         # Format everything appropriately for Avogadro
         # Start by passing back the original cjson, then add changes
         result = {"moleculeFormat": "cjson", "cjson": avo_input["cjson"]}
-        # Currently Avogadro ignores the energy result
-        result["message"] = (
-            f"Energy from GFN{config['method']}-xTB:\n"
-            + f"{str(round(energy_hartree, 7))} hartree\n"
-            + f"{str(round(energies['eV'], 7))} eV\n"
-            + f"{str(round(energies['kJ'], 7))} kJ/mol\n"
-            + f"{str(round(energies['kcal'], 7))} kcal/mol\n"
-        )
+        result["cjson"]["atoms"]["coords"] = cjson_geom["atoms"]["coords"]
         result["cjson"]["properties"]["totalEnergy"] = str(round(energies["eV"], 7))
+
+        # If the cjson contained frequencies or orbitals, remove them as they are no longer physical
+        for field in ["vibrations", "basisSet", "orbitals", "cube"]:
+            if field in result["cjson"]:
+                del result["cjson"][field]
+
+        # Save result
+        with open(calc_dir / "result.cjson", "w", encoding="utf-8") as save_file:
+            json.dump(result["cjson"], save_file, indent=2)
         # Pass back to Avogadro
         print(json.dumps(result))
